pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Евгений <prokofev_619@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
django-bootstrap5 = "^24.2"
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
